RepoSearchFragment と Adapter を分離し、Adapter の名前を変更
